renamed components to index.js, re-if in layout
