+++ src/App.js
@@ -1,7 +1,7 @@
 import "./App.css";
-import Header from "./components/Header/Header";
-import Layout from "./components/Layout/Layout";
-import Footer from "./components/Footer/Footer";
+import Header from "./components/Header";
+import Layout from "./components/Layout";
+import Footer from "./components/Footer";
 import bg1 from "./assets/bg1.jpg";
 import bg2 from "./assets/bg2.jpg";
 import bg3 from "./assets/bg3.jpg";
